fix: cast bf16 waveform to fp32 in MLX before numpy conversion
np.array(bf16_mlx_array, dtype=np.float32) fails via buffer protocol
because numpy has no bf16 equivalent. _save_waveform receives bf16
waveforms when the vocoder weights load in bf16 (q4/q8). Cast
explicitly with .astype(mx.float32) before handing to numpy.

+++ packages/ltx-pipelines-mlx/src/ltx_pipelines_mlx/ti2vid_one_stage.py
@@ -478,7 +478,7 @@ class TextToVideoPipeline:
             num_channels = 1
             wav = wav[:, None]  # (T, 1)
 
-        wav_np = np.array(wav, dtype=np.float32)
+        wav_np = np.array(wav.astype(mx.float32), dtype=np.float32)
         wav_np = np.clip(wav_np, -1.0, 1.0)
         wav_int16 = (wav_np * 32767).astype(np.int16)
